+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - `cli/cat`: Rename command `show` to `cat`
 - `cli/mv`: Rename command `move` to `mv`
 - `cli/env`: Rename command `project` to `env`
+- `cli/desk`: Rename command `board` to `desk`
 
 ### Fixed